requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
openai>=1.3.0
python-dotenv>=1.0.0
twilio>=8.10.0
python-multipart>=0.0.6
html2text>=2020.1.16
numpy>=1.24.0 